bug fixed

+++ app/src/main/java/com/ds/audio/video/screen/backgroundrecorder/Video_Record/services/Video_RecorderService.java
@@ -1,14 +1,11 @@
 package com.ds.audio.video.screen.backgroundrecorder.Video_Record.services;
 
-import static com.unity3d.services.core.properties.ClientProperties.getApplication;
-
 import android.app.AlarmManager;
 import android.app.Notification;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.app.Service;
-import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -31,14 +28,13 @@ import androidx.core.app.TaskStackBuilder;
 import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import com.ds.audio.video.screen.backgroundrecorder.BuildConfig;
-import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Activities.Video_ActivityPager;
 import com.ds.audio.video.screen.backgroundrecorder.CY_M_Define.CY_M_Conts;
+import com.ds.audio.video.screen.backgroundrecorder.R;
+import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Activities.Video_ActivityPager;
 import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Activities.Video_Save_Schedule_Activity;
-import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Fragment.Video_RecorderFragment;
 import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Helper.Video_CameraHelper;
 import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Helper.Video_FileHelper;
 import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Helper.Video_SharedPreHelper;
-import com.ds.audio.video.screen.backgroundrecorder.R;
 import com.ds.audio.video.screen.backgroundrecorder.Video_Record.Receiver.Video_AlarmReceiver;
 import com.ds.audio.video.screen.backgroundrecorder.roomdb.Video.WordRepository;
 import com.github.mylibrary.Notification.Ads.SharePrefUtils;
@@ -109,12 +105,16 @@ public class Video_RecorderService extends Service implements SurfaceHolder.Call
         NotificationCompat.Builder builder;
         super.onStartCommand(intent, i, i2);
         mRepository = new WordRepository(getApplication());
-        mRepository.deleteTimer(Double.parseDouble(SharePrefUtils.getString(CY_M_Conts.CURRENT_TIME, "")));
-        try {
-            Thread.sleep(10000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
+
+        if (!SharePrefUtils.getString(CY_M_Conts.CURRENT_TIME, "").equals("")) {
+            mRepository.deleteTimer(Double.parseDouble(SharePrefUtils.getString(CY_M_Conts.CURRENT_TIME, "")));
+            try {
+                Thread.sleep(10000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
         }
+
         if (Video_Save_Schedule_Activity.schedeluLisst.size() > 0) {
             Log.e("#TESTSCHEDULE", "2-->   " + Video_Save_Schedule_Activity.schedeluLisst.get(0).getTime());
             Intent intent1 = new Intent(this, Video_AlarmReceiver.class);
